feat(giraffe): test csvToTable
